Moved FhirAnonymousOperationType to Features.AnonymousOperations
Added Lazy value initialization for _fhirAnonymousOperationTypeList
