+++ src/Microsoft.Health.Fhir.Api/Features/AnonymousOperations/FhirAnonymousOperationType.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
-namespace Microsoft.Health.Fhir.ValueSets
+namespace Microsoft.Health.Fhir.Api.Features.AnonymousOperations
 {
     /// <summary>
     /// Value set for Fhir operations which do not require authorization
