+++ src/Microsoft.Health.Fhir.Api/Features/Audit/AuditHelper.cs
@@ -13,8 +13,8 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Health.Api.Features.Audit;
 using Microsoft.Health.Core.Features.Context;
 using Microsoft.Health.Core.Features.Security;
+using Microsoft.Health.Fhir.Api.Features.AnonymousOperations;
 using Microsoft.Health.Fhir.Core.Features.Context;
-using Microsoft.Health.Fhir.ValueSets;
 
 namespace Microsoft.Health.Fhir.Api.Features.Audit
 {
@@ -26,7 +26,7 @@ namespace Microsoft.Health.Fhir.Api.Features.Audit
         private readonly RequestContextAccessor<IFhirRequestContext> _fhirRequestContextAccessor;
         private readonly IAuditLogger _auditLogger;
         private readonly IAuditHeaderReader _auditHeaderReader;
-        private static IList<string> _fhirAnonymousOperationTypeList = new List<string>();
+        private static Lazy<IList<string>> _fhirAnonymousOperationTypeList;
 
         public AuditHelper(
             RequestContextAccessor<IFhirRequestContext> fhirRequestContextAccessor,
@@ -40,8 +40,10 @@ namespace Microsoft.Health.Fhir.Api.Features.Audit
             _fhirRequestContextAccessor = fhirRequestContextAccessor;
             _auditLogger = auditLogger;
             _auditHeaderReader = auditHeaderReader;
-            _fhirAnonymousOperationTypeList = GetConstants();
+            _fhirAnonymousOperationTypeList = new Lazy<IList<string>>(() => GetAnonymousOperations());
         }
+
+        public static IList<string> FhirAnonymousOperationTypeList => _fhirAnonymousOperationTypeList.Value;
 
         /// <inheritdoc />
         public void LogExecuting(HttpContext httpContext, IClaimsExtractor claimsExtractor)
@@ -78,7 +80,7 @@ namespace Microsoft.Health.Fhir.Api.Features.Audit
 
             // Audit the call if an audit event type is associated with the action.
             // Since AuditEventType holds value for both AuditEventType and FhirAnonymousOperationType ensure that we only log the AuditEventType
-            if (!string.IsNullOrEmpty(auditEventType) && !_fhirAnonymousOperationTypeList.Contains(auditEventType, StringComparer.OrdinalIgnoreCase))
+            if (!string.IsNullOrEmpty(auditEventType) && !FhirAnonymousOperationTypeList.Contains(auditEventType, StringComparer.OrdinalIgnoreCase))
             {
                 _auditLogger.LogAudit(
                     auditAction,
@@ -97,24 +99,22 @@ namespace Microsoft.Health.Fhir.Api.Features.Audit
         /// Return all the values of constants of the specified type
         /// </summary>
         /// <returns>List of constant values</returns>
-        public static IList<string> GetConstants()
+        private static IList<string> GetAnonymousOperations()
         {
-            if (_fhirAnonymousOperationTypeList.Count == 0)
-            {
-                FieldInfo[] fieldInfos = typeof(FhirAnonymousOperationType).GetFields(BindingFlags.Public | BindingFlags.Static | BindingFlags.FlattenHierarchy);
+            IList<string> anonymousOperations = new List<string>();
+            FieldInfo[] fieldInfos = typeof(FhirAnonymousOperationType).GetFields(BindingFlags.Public | BindingFlags.Static | BindingFlags.FlattenHierarchy);
 
-                // Go through the list and only pick out the constants
-                foreach (FieldInfo fi in fieldInfos)
+            // Go through the list and only pick out the constants
+            foreach (FieldInfo fi in fieldInfos)
+            {
+                if (fi.IsLiteral && !fi.IsInitOnly)
                 {
-                    if (fi.IsLiteral && !fi.IsInitOnly)
-                    {
-                        _fhirAnonymousOperationTypeList.Add(fi.Name);
-                    }
+                    anonymousOperations.Add(fi.Name);
                 }
             }
 
             // Return an array of FieldInfos
-            return _fhirAnonymousOperationTypeList;
+            return anonymousOperations;
         }
     }
 }
